+++ src/Router.tsx
@@ -4,8 +4,7 @@ import TimeLine from './pages/TimeLine/TimeLine';
 import Map from './pages/Map';
 import Explore from './pages/Explore/Explore';
 import UserProfile from './pages/UserProfile';
-import KakaoLogin from './pages/login/KakaoLogin';
-import GoogleLogin from './pages/login/GoogleLogin';
+import Login from './pages/login/Login';
 
 const Router = () => {
   const router = createBrowserRouter([
@@ -37,12 +36,12 @@ const Router = () => {
     {
       index: true,
       path: '/kakao/callback',
-      element: <KakaoLogin />,
+      element: <Login />,
     },
     {
       index: true,
       path: '/google/callback',
-      element: <GoogleLogin />,
+      element: <Login />,
     },
   ]);
